+++ inc/pid.h
@@ -2,7 +2,7 @@
 #define PID_H_
 
 void pid_setup_constants(double Kp_, double Ki_, double Kd_);
-void pid_update_reference(float reference);
+void pid_update_reference(double reference);
 double pid_control(double saida_medida);
 
 #endif
